+++ .eslintignore
@@ -1,2 +1,3 @@
 webpack.config.js
+bundle.js
