+++ js/debounce.js
@@ -1,17 +1,16 @@
 "use strict";
-(() => {
-  const INTERVAL = 300; // ms
 
-  let lastTimeOut = null;
+const INTERVAL = 300; // ms
 
-  window.debounce = (cb) => {
-    return (...parameters) => {
-      if (lastTimeOut) {
-        window.clearTimeout(lastTimeOut);
-      }
-      lastTimeOut = window.setTimeout(() => {
-        cb(...parameters);
-      }, INTERVAL);
-    };
+let lastTimeOut = null;
+
+window.debounce = (cb) => {
+  return (...parameters) => {
+    if (lastTimeOut) {
+      window.clearTimeout(lastTimeOut);
+    }
+    lastTimeOut = window.setTimeout(() => {
+      cb(...parameters);
+    }, INTERVAL);
   };
-})();
+};
