+++ package.json
@@ -12,12 +12,15 @@
   },
   "devDependencies": {
     "eslint": "7.1.0",
-    "eslint-config-htmlacademy": "0.6.0"
+    "eslint-config-htmlacademy": "0.6.0",
+    "webpack": "5.3.2",
+    "webpack-cli": "4.1.0"
   },
   "license": "MIT",
   "dependencies": {},
   "scripts": {
-    "test": "eslint js/"
+    "test": "eslint js/",
+    "build": "webpack --mode production"
   },
   "engines": {
     "node": "8.11"
